+++ docs/assets/navigation.js
@@ -1,1 +1,1 @@
-window.navigationData = "eJyVks1OwzAQhF8F7TmiUPEjcmspcAGBFG5VDlayaaymdmRvEBXquyMiEpzEjs3Vs/N57J3tFxB+EsTw2IiMuBQaIshKXuUKBcTbXs8UMsI3VJprQkEJSYUQQc2ohBiKzr6wDp6XdKgggj0XOcQ3V6doBPbg/gG5l6I9nGf9Ts0gD6h21kStMGNstPs1nRZgn3nHaGQCSw3c+7HGs1XFmUbXateyEfkoMh1r1Is/ZXjH8uLu9vJ6aVyzwYILXPUVGmAGoo/0hJQQo0mW7tznHzXvtbZGsk/52IkjWxKYrR16kTkvOKoJxBT9JKnYDn+2O+X0UgjF6g9zTipqAKzddHAePlDYKa0SxHhut2n7VUMMIjka4+9Jekq/AXMFzbs="
+window.navigationData = "eJyNkstOw0AMRX8FeV1RqHiI7NqCYAECKbCqsogSp7GazkQzDqJC/XdESso0mUe2c6+Prz1efQPjF0MES6kQJpCVVOUKBUSro5QpTBnfUGnSjIJjPnjrlEuIoGhExiSFnlqN5yVvK5jAhkQO0c3VftIDB3B2SGJgnqTcaEf2Rrv5neZJ2FmWUrQvHsqfxZv0g6kiJnSlNSb2dBy6PBNsUa2t07eCN+37rsazeUWpdgZeyEbkvQXzrkY9/VdOe8wu7m4vr2dGm3ssSOD8EKuPORFDpEfkmFMeZOneQ/W9y32trZHsrhA7dmSLR2ZrTS8yp4JQDSCmGCZJla7x93eHnKM0hmKtH1c5OG8DYD1qB+fhE4Wd0iqjGM/tb9q2aoijSI6LCd9Jsk9+AAt34Z8="
