+++ src/lib/hooks/useStore.ts
@@ -1,6 +1,6 @@
 import { useRef, useSyncExternalStore } from "react";
-import { StateModifier, Store } from "./createStore";
-import { deeplyEquals } from "./deeplyEquals";
+import { StateModifier, Store } from "../core/createStore";
+import { deeplyEquals } from "../utils/deeplyEquals";
 
 export type BoundStore<
   TState extends object,
@@ -52,6 +52,8 @@ export type BoundStore<
  * If the `select` function is provided, an equality check is performed. This has some caveats:
  * - For optimal performance, return a direct reference to the state. (e.g. `state.count`)
  * - If you return an object literal, it should only contain direct references to the state. (e.g. `{ count: state.count }`)
+ *
+ * @group Hooks
  */
 export const useStore = <
   TState extends object,
